OP_CHECKSIG

+++ bitcoin.sql
@@ -126,7 +126,7 @@ CREATE TABLE transactions (
 CREATE TABLE outputs (
     output_id INT NOT NULL DEFAULT NEXTVAL('outputs_output_id_sequence') PRIMARY KEY,
     parent_id INT NOT NULL,
-    index_in_parent INT NOT NULL,
+    index_in_parent BIGINT NOT NULL,
     script_id INT NOT NULL,
     value amount_type NOT NULL,
     output_type output_transaction_type NOT NULL,
